Create unit test for sparse arrays
